Refactor: Move the sound functions to Vehicle.Sound.cpp
Part of #25951

This moves vehicle sound logic like UpdateSound, scream and friction sound handling, lift hill sounds and Claxon to Vehicle.Sound.cpp. Also, it promotes SoundIdVolume struct definition from Vehicle.cpp to Vehicle.h to allow shared access across translation units. Music to my ears. Except.. not really.
